requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
asgiref==3.7.2
    # via django
django==4.2.9
    # via
    #   -r requirements.in
    #   djangorestframework
django-behaviors==0.5.1
    # via -r requirements.in
django-environ==0.11.2
    # via -r requirements.in
djangorestframework==3.14.0
    # via -r requirements.in
psycopg2-binary==2.9.9
    # via -r requirements.in
pytz==2023.3.post1
    # via djangorestframework
sqlparse==0.4.4
    # via django
